Use preferred types

+++ src/rawsource.h
@@ -24,7 +24,7 @@ struct RawSource {
     uint32_t uses;
 
     // Read num_bytes into buffer from fd starting at offset
-    ssize_t offset_read(void* buffer, size_t num_bytes, off_t offset) const {
+    int64_t offset_read(void* buffer, uint64_t num_bytes, uint64_t offset) const {
 #ifdef _WIN32
         // The POSIX version (pread) does not change the file cursor, this does.
         // Furthermore, this is not thread-safe!
